Fixed #321
