+++ src/main/java/com/box/sdk/BoxGroupIterator.java
@@ -1,23 +1,12 @@
 package com.box.sdk;
 
 import java.net.URL;
-import java.util.Iterator;
 
 import com.eclipsesource.json.JsonObject;
 
-class BoxGroupIterator implements Iterator<BoxGroup.Info> {
-    private static final long LIMIT = 1000;
-
-    private final BoxAPIConnection api;
-    private final JSONIterator jsonIterator;
-
+class BoxGroupIterator extends BoxIteratorBase<BoxGroup.Info> {
     BoxGroupIterator(BoxAPIConnection api, URL url) {
-        this.api = api;
-        this.jsonIterator = new JSONIterator(api, url, LIMIT);
-    }
-
-    public boolean hasNext() {
-        return this.jsonIterator.hasNext();
+        super(api, url);
     }
 
     public BoxGroup.Info next() {
@@ -26,9 +15,5 @@ class BoxGroupIterator implements Iterator<BoxGroup.Info> {
 
         BoxGroup group = new BoxGroup(this.api, id);
         return group.new Info(nextJSONObject);
-    }
-
-    public void remove() {
-        throw new UnsupportedOperationException();
     }
 }
