+++ src/main/java/com/box/sdk/BoxUserIterator.java
@@ -1,23 +1,12 @@
 package com.box.sdk;
 
 import java.net.URL;
-import java.util.Iterator;
 
 import com.eclipsesource.json.JsonObject;
 
-class BoxUserIterator implements Iterator<BoxUser.Info> {
-    private static final long LIMIT = 1000;
-
-    private final BoxAPIConnection api;
-    private final JSONIterator jsonIterator;
-
+class BoxUserIterator extends BoxIteratorBase<BoxUser.Info> {
     BoxUserIterator(BoxAPIConnection api, URL url) {
-        this.api = api;
-        this.jsonIterator = new JSONIterator(api, url, LIMIT);
-    }
-
-    public boolean hasNext() {
-        return this.jsonIterator.hasNext();
+        super(api, url);
     }
 
     public BoxUser.Info next() {
@@ -26,9 +15,5 @@ class BoxUserIterator implements Iterator<BoxUser.Info> {
 
         BoxUser user = new BoxUser(this.api, id);
         return user.new Info(nextJSONObject);
-    }
-
-    public void remove() {
-        throw new UnsupportedOperationException();
     }
 }
